mock polls transitions better
